+++ src/main/java/com/chanthai/phoneshop/service/BrandService.java
@@ -8,8 +8,8 @@ import java.util.Map;
 
 public interface BrandService {
     Brand create(Brand brand);
-    Brand getById(Integer id);
-    Brand update(Integer id, Brand brandUpdate);
+    Brand getById(Long id);
+    Brand update(Long id, Brand brandUpdate);
     List<Brand> getBrands(String name);
     //List<Brand> getBrands(Map<String,String> params);
     Page<Brand> getBrands(Map<String,String> params);
